Fact Decorator Rough Sketch
This would introduce the concept of a FactDecorator to match
the idea of a operator decorator. Fact Decorators could be used to do things like  getting keysOf, valuesOf, or sizeOf facts.
Additionally in order to Remove JSONPathPlus as such a dependency we could use the fact decorator pattern, even if we wanted to ship our own decorator that depended on JSON path we could move it to a peer dependency.